refactor: move profile helpers and constants to lib/util/profile
Extracts PROFILE_ID_PATTERN, ASSET_FILENAME_PATTERN, validateProfileSubmission
and validateProfileFilesystem out of app.js and into lib/util/profile so the
HTTP layer holds only middleware and routes. Drops the PROFILE_ID_PATTERN
duplicate that the regex previously had between the schema validator and the
app module. The multer middleware declarations move up next to the other
express middleware so app.js reads as requires, middleware, routes.
