+++ app.js
@@ -57,6 +57,29 @@ app.locals.dev = process.env.NODE_ENV !== "production";
 app.use("/static", express.static(path.join(__dirname, "static")));
 app.use(bodyParser.urlencoded({ extended: true }));
 
+// configures the multer middleware that parses the multipart
+// payload of the profile form, accepting only text fields since
+// background images are now managed by the dedicated asset
+// endpoints under `/profiles/assets`
+const profileUpload = multer().none();
+
+// configures the multer middleware that handles the multipart
+// upload of a single PNG asset, keeping it in memory so the
+// magic bytes can be validated before the file is written
+const assetUpload = multer({
+    storage: multer.memoryStorage(),
+    limits: { fileSize: 5 * 1024 * 1024, files: 1 }
+}).single("file");
+
+// configures the multer middleware that handles the multipart
+// upload of a single profiles bundle zip, keeping it in memory
+// so the archive can be unpacked before the on disk profiles
+// directory is replaced
+const bundleUpload = multer({
+    storage: multer.memoryStorage(),
+    limits: { fileSize: 50 * 1024 * 1024, files: 1 }
+}).single("file");
+
 app.get("/", (req, res, next) => {
     // forwards the bare root URL to the user's preferred home
     // landing page, defaulting to the classic gateway when no
@@ -312,125 +335,6 @@ app.get("/config", (req, res, next) => {
     res.json(req.session.config || {});
 });
 
-// configures the multer middleware that parses the multipart
-// payload of the profile form, accepting only text fields since
-// background images are now managed by the dedicated asset
-// endpoints under `/profiles/assets`
-const profileUpload = multer().none();
-
-// configures the multer middleware that handles the multipart
-// upload of a single PNG asset, keeping it in memory so the
-// magic bytes can be validated before the file is written
-const assetUpload = multer({
-    storage: multer.memoryStorage(),
-    limits: { fileSize: 5 * 1024 * 1024, files: 1 }
-}).single("file");
-
-// configures the multer middleware that handles the multipart
-// upload of a single profiles bundle zip, keeping it in memory
-// so the archive can be unpacked before the on disk profiles
-// directory is replaced
-const bundleUpload = multer({
-    storage: multer.memoryStorage(),
-    limits: { fileSize: 50 * 1024 * 1024, files: 1 }
-}).single("file");
-
-// matches the validated filename pattern accepted by the asset
-// endpoints so arbitrary paths cannot be crafted via the URL or
-// form fields when reading, writing, or deleting on disk
-const ASSET_FILENAME_PATTERN = /^[a-z0-9]+(?:-[a-z0-9]+)*\.png$/;
-
-// validates the profile and inspirations JSON payloads coming
-// from the upload form, returning the parsed objects together
-// with the accumulated error messages so callers can render or
-// persist based on the outcome
-const validateProfileSubmission = function(profileText, inspirationsText) {
-    const errors = [];
-    let profile = null;
-    let inspirations = null;
-
-    if (!profileText || !profileText.trim()) {
-        errors.push("profile JSON is required");
-    } else {
-        try {
-            profile = JSON.parse(profileText);
-        } catch (err) {
-            errors.push("profile JSON is not valid JSON: " + err.message);
-        }
-    }
-    if (profile) {
-        const profileErrors = lib.validateProfile(profile);
-        for (const message of profileErrors) errors.push("profile: " + message);
-    }
-
-    if (inspirationsText && inspirationsText.trim()) {
-        try {
-            inspirations = JSON.parse(inspirationsText);
-        } catch (err) {
-            errors.push("inspirations JSON is not valid JSON: " + err.message);
-        }
-        if (inspirations) {
-            const inspirationErrors = lib.validateInspirations(inspirations);
-            for (const message of inspirationErrors) errors.push("inspirations: " + message);
-        }
-    }
-
-    return { profile: profile, inspirations: inspirations, errors: errors };
-};
-
-// matches the validated identifier pattern accepted by the
-// profile schema so that arbitrary filesystem paths cannot be
-// crafted from a tampered query string or form field
-const PROFILE_ID_PATTERN = /^[a-z0-9]+(?:-[a-z0-9]+)*$/;
-
-// validates the on disk consequences of the submitted profile,
-// enforcing different rules for create and edit mode so the user
-// can confidently rename in place without accidentally clobbering
-// another template or losing track of the original
-const validateProfileFilesystem = async function(profile, editTarget, directoryPath) {
-    const errors = [];
-    if (!profile || !profile.id) return errors;
-
-    if (editTarget) {
-        // edit mode requires the original template to still exist
-        // on disk so a stale browser tab cannot silently resurrect
-        // a profile that was deleted from another session
-        const targetPath = path.join(directoryPath, `${editTarget}.json`);
-        try {
-            await fs.access(targetPath);
-        } catch (err) {
-            errors.push(`edited template "${editTarget}" no longer exists`);
-            return errors;
-        }
-
-        // edit mode allows renaming the profile id but refuses to
-        // overwrite a different existing template so the rename
-        // path cannot be used to clobber unrelated entries
-        if (profile.id !== editTarget) {
-            const renamePath = path.join(directoryPath, `${profile.id}.json`);
-            try {
-                await fs.access(renamePath);
-                errors.push(`profile with id "${profile.id}" already exists`);
-            } catch (err) {
-                // target id is free, the rename is allowed to proceed
-            }
-        }
-        return errors;
-    }
-
-    // create mode refuses to overwrite any existing profile so a
-    // mistaken upload cannot replace a default template; the user
-    // must pick a different id to retry
-    const profilePath = path.join(directoryPath, `${profile.id}.json`);
-    try {
-        await fs.access(profilePath);
-        errors.push(`profile with id "${profile.id}" already exists`);
-    } catch (err) {
-        // file does not exist, which is the expected state
-    }
-    return errors;
-};
-
 app.get("/profiles/manager", (req, res, next) => {
     async function clojure() {
         const fullscreen = req.query.fullscreen === "1";
@@ -446,7 +350,7 @@ app.get("/profiles/manager", (req, res, next) => {
         let profileJson = "";
         let inspirationsJson = "";
         let editTarget = "";
-        if (editId && PROFILE_ID_PATTERN.test(editId)) {
+        if (editId && lib.PROFILE_ID_PATTERN.test(editId)) {
             const directoryPath = path.join(__dirname, "static", "profiles");
             try {
                 const profileContent = await fs.readFile(
@@ -490,12 +394,12 @@ app.get("/profiles/manager", (req, res, next) => {
 app.post("/profiles/validate", profileUpload, (req, res, next) => {
     async function clojure() {
         const editTarget = typeof req.body.edit_target === "string" ? req.body.edit_target : "";
-        const { profile, errors } = validateProfileSubmission(
+        const { profile, errors } = lib.validateProfileSubmission(
             req.body.profile_json,
             req.body.inspirations_json
         );
         const directoryPath = path.join(__dirname, "static", "profiles");
-        const filesystemErrors = await validateProfileFilesystem(
+        const filesystemErrors = await lib.validateProfileFilesystem(
             profile,
             editTarget,
             directoryPath
@@ -511,7 +415,7 @@ app.post("/profiles", profileUpload, (req, res, next) => {
         const profileText = req.body.profile_json || "";
         const inspirationsText = req.body.inspirations_json || "";
         const editTarget = typeof req.body.edit_target === "string" ? req.body.edit_target : "";
-        const { profile, inspirations, errors } = validateProfileSubmission(
+        const { profile, inspirations, errors } = lib.validateProfileSubmission(
             profileText,
             inspirationsText
         );
@@ -521,7 +425,7 @@ app.post("/profiles", profileUpload, (req, res, next) => {
         // renames without clobbering unrelated templates
         const directoryPath = path.join(__dirname, "static", "profiles");
         if (errors.length === 0) {
-            const filesystemErrors = await validateProfileFilesystem(
+            const filesystemErrors = await lib.validateProfileFilesystem(
                 profile,
                 editTarget,
                 directoryPath
@@ -571,7 +475,7 @@ app.post("/profiles", profileUpload, (req, res, next) => {
 app.post("/profiles/:id/enabled", profileUpload, (req, res, next) => {
     async function clojure() {
         const id = req.params.id;
-        if (!PROFILE_ID_PATTERN.test(id)) {
+        if (!lib.PROFILE_ID_PATTERN.test(id)) {
             res.status(400).json({ error: "invalid profile id" });
             return;
         }
@@ -610,7 +514,7 @@ app.post("/profiles/:id/enabled", profileUpload, (req, res, next) => {
 app.post("/profiles/:id/delete", (req, res, next) => {
     async function clojure() {
         const id = req.params.id;
-        if (!PROFILE_ID_PATTERN.test(id)) {
+        if (!lib.PROFILE_ID_PATTERN.test(id)) {
             res.status(400).json({ error: "invalid profile id" });
             return;
         }
@@ -649,7 +553,7 @@ app.get("/profiles/assets", (req, res, next) => {
     async function clojure() {
         const directoryPath = path.join(__dirname, "static", "profiles");
         const files = await fs.readdir(directoryPath);
-        const assets = files.filter(file => ASSET_FILENAME_PATTERN.test(file)).sort();
+        const assets = files.filter(file => lib.ASSET_FILENAME_PATTERN.test(file)).sort();
         res.json({ assets: assets });
     }
     clojure().catch(next);
@@ -665,7 +569,7 @@ app.post("/profiles/assets", assetUpload, (req, res, next) => {
         const filename = typeof req.body.filename === "string" ? req.body.filename.trim() : "";
         if (!filename) {
             errors.push("filename is required");
-        } else if (!ASSET_FILENAME_PATTERN.test(filename)) {
+        } else if (!lib.ASSET_FILENAME_PATTERN.test(filename)) {
             errors.push(
                 "filename must match pattern: lowercase alphanumeric with hyphens and a .png extension"
             );
@@ -715,7 +619,7 @@ app.post("/profiles/assets", assetUpload, (req, res, next) => {
 app.post("/profiles/assets/:filename/delete", (req, res, next) => {
     async function clojure() {
         const filename = req.params.filename;
-        if (!ASSET_FILENAME_PATTERN.test(filename)) {
+        if (!lib.ASSET_FILENAME_PATTERN.test(filename)) {
             res.status(400).json({ error: "invalid asset filename" });
             return;
         }
