+++ lib/util/profile.js
@@ -1,3 +1,6 @@
+const fs = require("fs/promises");
+const path = require("path");
+
 const VALID_UNITS = ["mm", "px"];
 const VALID_ORIENTATIONS = ["portrait", "landscape"];
 const VALID_SHAPES = ["rectangle", "circle", "heart"];
@@ -5,7 +8,15 @@ const VALID_FONT_SIZE_MODES = ["manual", "automatic"];
 const VALID_TEXT_ALIGNS = ["left", "center", "right"];
 const VALID_VERTICAL_ALIGNS = ["top", "middle", "bottom"];
 
+// matches the validated identifier pattern accepted by the
+// profile schema so that arbitrary filesystem paths cannot be
+// crafted from a tampered query string or form field
 const PROFILE_ID_PATTERN = /^[a-z0-9]+(?:-[a-z0-9]+)*$/;
+
+// matches the validated filename pattern accepted by the asset
+// endpoints so arbitrary paths cannot be crafted via the URL or
+// form fields when reading, writing, or deleting on disk
+const ASSET_FILENAME_PATTERN = /^[a-z0-9]+(?:-[a-z0-9]+)*\.png$/;
 
 /**
  * Validates that the provided value is a positive number
@@ -506,7 +517,112 @@ const validateProfile = profile => {
     return errors;
 };
 
+/**
+ * Validates the profile and inspirations JSON payloads coming
+ * from the upload form, returning the parsed objects together
+ * with the accumulated error messages so callers can render or
+ * persist based on the outcome.
+ *
+ * @param {String} profileText The profile JSON payload as text.
+ * @param {String} inspirationsText The optional inspirations JSON
+ *     payload as text, ignored when empty.
+ * @returns {Object} An object with parsed profile, inspirations,
+ *     and an errors array (empty if valid).
+ */
+const validateProfileSubmission = function(profileText, inspirationsText) {
+    const errors = [];
+    let profile = null;
+    let inspirations = null;
+
+    if (!profileText || !profileText.trim()) {
+        errors.push("profile JSON is required");
+    } else {
+        try {
+            profile = JSON.parse(profileText);
+        } catch (err) {
+            errors.push("profile JSON is not valid JSON: " + err.message);
+        }
+    }
+    if (profile) {
+        const profileErrors = validateProfile(profile);
+        for (const message of profileErrors) errors.push("profile: " + message);
+    }
+
+    if (inspirationsText && inspirationsText.trim()) {
+        try {
+            inspirations = JSON.parse(inspirationsText);
+        } catch (err) {
+            errors.push("inspirations JSON is not valid JSON: " + err.message);
+        }
+        if (inspirations) {
+            const inspirationErrors = validateInspirations(inspirations);
+            for (const message of inspirationErrors) errors.push("inspirations: " + message);
+        }
+    }
+
+    return { profile: profile, inspirations: inspirations, errors: errors };
+};
+
+/**
+ * Validates the on disk consequences of the submitted profile,
+ * enforcing different rules for create and edit mode so the user
+ * can confidently rename in place without accidentally clobbering
+ * another template or losing track of the original.
+ *
+ * @param {Object} profile The parsed profile object.
+ * @param {String} editTarget The id of the profile being edited,
+ *     or an empty value for create mode.
+ * @param {String} directoryPath The absolute path of the on disk
+ *     profiles directory.
+ * @returns {Array} An array of error messages (empty if valid).
+ */
+const validateProfileFilesystem = async function(profile, editTarget, directoryPath) {
+    const errors = [];
+    if (!profile || !profile.id) return errors;
+
+    if (editTarget) {
+        // edit mode requires the original template to still exist
+        // on disk so a stale browser tab cannot silently resurrect
+        // a profile that was deleted from another session
+        const targetPath = path.join(directoryPath, `${editTarget}.json`);
+        try {
+            await fs.access(targetPath);
+        } catch (err) {
+            errors.push(`edited template "${editTarget}" no longer exists`);
+            return errors;
+        }
+
+        // edit mode allows renaming the profile id but refuses to
+        // overwrite a different existing template so the rename
+        // path cannot be used to clobber unrelated entries
+        if (profile.id !== editTarget) {
+            const renamePath = path.join(directoryPath, `${profile.id}.json`);
+            try {
+                await fs.access(renamePath);
+                errors.push(`profile with id "${profile.id}" already exists`);
+            } catch (err) {
+                // target id is free, the rename is allowed to proceed
+            }
+        }
+        return errors;
+    }
+
+    // create mode refuses to overwrite any existing profile so a
+    // mistaken upload cannot replace a default template; the user
+    // must pick a different id to retry
+    const profilePath = path.join(directoryPath, `${profile.id}.json`);
+    try {
+        await fs.access(profilePath);
+        errors.push(`profile with id "${profile.id}" already exists`);
+    } catch (err) {
+        // file does not exist, which is the expected state
+    }
+    return errors;
+};
+
 module.exports = {
+    PROFILE_ID_PATTERN: PROFILE_ID_PATTERN,
+    ASSET_FILENAME_PATTERN: ASSET_FILENAME_PATTERN,
     validateProfile: validateProfile,
     validatePadding: validatePadding,
     validateExtraPadding: validateExtraPadding,
@@ -519,5 +635,7 @@ module.exports = {
     validateText: validateText,
     validateMetadata: validateMetadata,
     validateInspiration: validateInspiration,
-    validateInspirations: validateInspirations
+    validateInspirations: validateInspirations,
+    validateProfileSubmission: validateProfileSubmission,
+    validateProfileFilesystem: validateProfileFilesystem
 };
